+++ src/main/java/com/flowiee/pms/repository/product/ProductDescriptionRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductDescriptionRepository extends JpaRepository<ProductDescription, Integer> {
+public interface ProductDescriptionRepository extends JpaRepository<ProductDescription, Long> {
 }
