+++ src/main/java/com/flowiee/pms/repository/product/ProductHistoryRepository.java
@@ -9,10 +9,10 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ProductHistoryRepository extends JpaRepository<ProductHistory, Integer> {
+public interface ProductHistoryRepository extends JpaRepository<ProductHistory, Long> {
     @Query("from ProductHistory p where p.product.id=:productId")
-    List<ProductHistory> findByProductId(@Param("productId") Integer productId);
+    List<ProductHistory> findByProductId(@Param("productId") Long productId);
 
     @Query("from ProductHistory p where p.productDetail.id=:productVariantId and p.field=:field order by p.id desc")
-    List<ProductHistory> findHistoryChangeOfProductDetail(@Param("productVariantId") Integer productVariantId, @Param("field") String field);
+    List<ProductHistory> findHistoryChangeOfProductDetail(@Param("productVariantId") Long productVariantId, @Param("field") String field);
 }
